+++ docs/bookmarklets/opena11y-example.js
@@ -1121,15 +1121,32 @@
   const debug$12 = new DebugLogging('linkResults', false);
   debug$12.flag = false;
 
-  const allowedExtensions = [
-    'ai',    // Adobe Illustrator file'
-    'aif',   // AIF audio file
-    'bmp',   // Bitmap image
-    'cda',   // CD (Compact Disc) audio track file
+  const pdfExtensions = [
+    'pdf'   // Protable document format
+  ];
+
+  const docExtensions = [
     'csv',   // Comma separated spaces
     'doc',   // Microsoft Word Document
     'docx',
     'epub',  // Electronic book format
+    'odg',   // Open Document Graphics
+    'odp',   // Open Document Presentations
+    'ods',   // Open Document Spreadsheet
+    'odt',   // Open Document Text
+    'ppt',   // Microsoft Powerpoint document
+    'pptx',
+    'rtf',   // Rich text format
+    'rtfd',
+    'txt',   // Text file
+    'xls',   // Microsoft Spreadsheet document
+    'xlsx'
+  ];
+
+  const mediaExtensions = [
+    'aif',   // AIF audio file
+    'bmp',   // Bitmap image
+    'cda',   // CD (Compact Disc) audio track file
     'ico',   // Icon file
     'jpeg',  // JPEG (Joint Photographic Experts Group) image
     'jpg',
@@ -1139,30 +1156,21 @@
     'mp3',   // MP3 audio file
     'mpa',   // MPEG-2 audio file
     'png',   // PNG (Portable Network Graphics) image
-    'odg',   // Open Document Graphics
-    'odp',   // Open Document Presentations
-    'ods',   // Open Document Spreadsheet
-    'odt',   // Open Document Text
     'ogg',   // Ogg Vorbis audio file
-    'pdf',   // Protable document format
-    'ppt',   // Microsoft Powerpoint document
-    'pptx',
     'ps',    // PostScript file
     'psd',   // PSD (Photoshop document) image
-    'rtf',   // Rich text format
-    'rtfd',
     'svg',   // Scalable Vector Graphics file
-    'tar',   // Linux / Unix tarball file archive
-    'gz',    // Tarball compressed file'
     'tif',   // TIFF (Tagged Image File Format) image
     'tiff',
-    'txt',   // Text file
     'wav',   // WAV file
     'webp',  // WebP image.
     'wma',   // WMA (Windows Media Audio) audio file
-    'wpl',   // Windows Media Player playlist
-    'xls',   // Microsoft Spreadsheet document
-    'xlsx',
+    'wpl'   // Windows Media Player playlist
+  ];
+
+  const zipExtensions = [
+    'tar',   // Linux / Unix tarball file archive
+    'gz',    // Tarball compressed file'
     'zip',   // Comprerssed file format
     '7z'     // 7-Zip compressed file
   ];
@@ -1186,47 +1194,62 @@
 
     update(domCache, url) {
 
-      const parsedURL =  URL.parse(url);
-
-      debug$12.flag && debug$12.log(`[parsedURL][hostname]: ${parsedURL.hostname}`);
-      debug$12.flag && debug$12.log(`[parsedURL][pathname]: ${parsedURL.pathname}`);
+      const parsedUrl   = URL.parse(url);
+      const partsUrl    = parsedUrl && parsedUrl.hostname ?
+                          parsedUrl.hostname.split('.') :
+                          '';
+      const partsUrlLen = parsedUrl && parsedUrl.hostname ?
+                          partsUrl.length :
+                          0;
 
       this.linkData = [];
 
       domCache.linkInfo.allLinkDomElements.forEach( de => {
 
-        const parsedHREF = URL.parse(de.node.href);
+        const parsedHref = URL.parse(de.node.href);
+        const partsHref  = parsedHref && parsedHref.hostname ?
+                           parsedHref.hostname.split('.') :
+                           '';
+        const partsHrefLen = parsedHref && parsedHref.hostname ?
+                             partsHref.length :
+                             0;
 
-        debug$12.flag && debug$12.log(`[parsedHREF]: ${parsedHREF}`);
+        if (parsedHref && parsedHref.hostname) {
 
-        if (parsedHREF) {
-          debug$12.flag && debug$12.log(`[parsedHREF][    href]: ${parsedHREF.href}`);
-          debug$12.flag && debug$12.log(`[parsedHREF][hostname]: ${parsedHREF.hostname} (${parsedURL.hostname === parsedHREF.hostname})`);
-          debug$12.flag && debug$12.log(`[parsedHREF][pathname]: ${parsedHREF.pathname}`);
-          debug$12.flag && debug$12.log(`[parsedHREF][    hash]: ${parsedHREF.hash}`);
-          debug$12.flag && debug$12.log(`[parsedHREF][  origin]: ${parsedHREF.origin}`);
+          const sameHostname = parsedUrl.hostname === parsedHref.hostname;
+          const sameDomain   = (partsUrlLen > 1 && partsHrefLen > 1) ?
+                               (partsUrl[partsUrlLen-1] === partsHref[partsHrefLen-1]) &&
+                               (partsUrl[partsUrlLen-2] === partsHref[partsHrefLen-2]) :
+                               false;
 
-          const sameHostname = parsedURL.hostname === parsedHREF.hostname;
-          const samePathname = parsedURL.pathname === parsedHREF.pathname;
+          const samePathname = parsedUrl.pathname === parsedHref.pathname;
 
-          const periodIndex   = parsedURL.pathname.lastIndexOf('.');
-          const extension     = periodIndex > 0 &&
-                                ((parsedURL.pathname - periodIndex) < 5) ?
-                                parsedURL.pathname.substring(periodIndex).trim().toLowerCase() :
+          const periodIndex   = parsedHref.pathname.lastIndexOf('.');
+          const extension     = periodIndex > 0 ?
+                                parsedHref.pathname.substring(periodIndex+1).trim().toLowerCase() :
                                 '';
 
-          const allowedExt    = allowedExtensions.includes(extension) ?
-                                extension :
+          const extensionType = pdfExtensions.includes(extension) ?
+                                'pdf' :
+                                docExtensions.includes(extension) ?
+                                'doc' :
+                                mediaExtensions.includes(extension) ?
+                                'media' :
+                                zipExtensions.includes(extension) ?
+                                'zip' :
                                 '';
 
           const dataItem = {
             url:               de.node.href,
             name:              cleanName(de.accName.name),
+            desc:              cleanName(de.accDescription.name),
             ordinalPosition:   de.ordinalPosition,
             isInternal:        sameHostname && samePathname,
-            isExternal:        !sameHostname,
-            isSameDomain:      sameHostname && !samePathname,
-            extension:         allowedExt,
+            isExternal:        !sameDomain,
+            isSameDomain:      sameDomain,
+            isSameSubDomain:   sameHostname,
+            extension:         extension,
+            extensionType:     extensionType,
             isVisibleOnScreen: de.visibility.isVisibleOnScreen,
             isVisibleToAT:     de.visibility.isVisibleToAT
           };
@@ -12666,6 +12689,8 @@
         debug$$.tag(elementNode);
       }
 
+      this.hasTextNodes = this.getHasTextNodes(elementNode);
+
       this.opacity            = this.normalizeOpacity(style, parentColorContrast);
 
       this.backgroundColorElem = style.getPropertyValue("background-color");
@@ -12702,6 +12727,30 @@
         debug$$.log(`[ Family/Size/Weight/isLarge]: "${this.fontFamily}"/${this.fontSize}/${this.fontWeight}/${this.isLargeFont}`);
         debug$$.color(`[   CCR for Color/Background]: ${this.colorContrastRatio} for #${this.colorHex}/#${this.backgroundColorHex}`, this.color, this.backgroundColor);
       }
+    }
+
+    /**
+     * @method getHasTextNodes
+     *
+     * @desc Returns true if the element node has text content, otherwise false
+     *       Analyzes node for text nodes
+     *
+     * @param {Object}  elemNode  - DOM element node
+     *
+     * @return {Number}  see @desc
+     */
+
+    getHasTextNodes (elemNode) {
+      let text = '';
+      let childNodes = elemNode.childNodes;
+
+      for (let i = 0; i < childNodes.length; i++) {
+        let node = childNodes[i];
+        if (node.nodeType === Node.TEXT_NODE) {
+          text += node.textContent.trim().replace(' ', '');
+        }
+      }
+      return text.length > 0;
     }
 
     /**
@@ -15005,6 +15054,7 @@
       this.isVisibilityHidden = this.normalizeVisibility (style, parentVisibility);
       this.isSmallHeight      = this.normalizeHeight(style, parentVisibility);
       this.isSmallFont        = this.getFontSize(style);
+      this.isInClosedDetails  = this.normalizeInClosedDetails(elementNode, parentVisibility);
 
       // Set default values for visibility
       this.isVisibleOnScreen = true;
@@ -15012,7 +15062,8 @@
 
       if (this.isHidden ||
           this.isDisplayNone ||
-          this.isVisibilityHidden) {
+          this.isVisibilityHidden ||
+          this.isInClosedDetails) {
 
         if (tagName !== 'area') {
           this.isVisibleOnScreen = false;
@@ -15038,9 +15089,31 @@
         debug$Y.log('[isVisibilityHidden]: ' + this.isVisibilityHidden);
         debug$Y.log('[     isSmallHeight]: ' + this.isSmallHeight);
         debug$Y.log('[       isSmallFont]: ' + this.isSmallFont);
+        debug$Y.log('[ isInClosedDetails]: ' + this.isInClosedDetails);
         debug$Y.log('[ isVisibleOnScreen]: ' + this.isVisibleOnScreen);
         debug$Y.log('[     isVisibleToAT]: ' + this.isVisibleToAT);
       }
+    }
+
+    /**
+     * @method normalizeInClosedDetails
+     *
+     * @desc Returns true if element is in a closed details eleemnt
+     *
+     * @param {Object}  node              - dom element node
+     * @param {Object}  parentVisibility  - Computed visibility information for parent
+     *                                      DomElement
+     *
+     * @return {Boolean} see @desc
+     */
+
+    normalizeInClosedDetails (node, parentVisibility) {
+      if (node.parentNode.tagName === 'DETAILS' &&
+          node.tagName !== 'SUMMARY' &&
+          !node.parentNode.open) {
+        return true;
+      }
+      return parentVisibility.isInClosedDetails;
     }
 
     /**
@@ -15643,7 +15716,7 @@
   *   @returns  {Boolean} see @desc
   */
 
-  function isDisplayNone (node) {
+  function isDisplayNone (node, psuedo=null) {
 
     if (!node) {
       return false;
@@ -15667,7 +15740,7 @@
         }
       }
 
-      const style = window.getComputedStyle(node, null);
+      const style = window.getComputedStyle(node, psuedo);
 
       const display = style.getPropertyValue("display");
 
@@ -15689,7 +15762,7 @@
   *   @return  see @desc
   */
 
-  function isVisibilityHidden(node) {
+  function isVisibilityHidden(node, psuedo=null) {
 
     if (!node) {
       return false;
@@ -15700,7 +15773,7 @@
     }
 
     if (node.nodeType === Node.ELEMENT_NODE) {
-      const style = window.getComputedStyle(node, null);
+      const style = window.getComputedStyle(node, psuedo);
 
       const visibility = style.getPropertyValue("visibility");
       if (visibility) {
@@ -15886,9 +15959,37 @@
   *   values, the result cannot and will not be equal to 'none'.
   */
   function addCssGeneratedContent (element, contents) {
-    let result = contents,
-        prefix = getComputedStyle(element, ':before').content,
-        suffix = getComputedStyle(element, ':after').content;
+
+    function isVisible (style) {
+
+      let flag = true;
+
+      const display = style.getPropertyValue("display");
+      if (display) {
+        flag = flag && display !== 'none';
+      }
+
+      const visibility = style.getPropertyValue("visibility");
+      if (visibility) {
+        flag = flag && (visibility !== 'hidden') && (visibility !== 'collapse');
+      }
+      return flag;
+    }
+
+    let result = contents;
+    const styleBefore = getComputedStyle(element, ':before');
+    const styleAfter  = getComputedStyle(element, ':after');
+
+    const beforeVisible = isVisible(styleBefore);
+    const afterVisible  = isVisible(styleAfter);
+
+    const prefix = beforeVisible ?
+                   styleBefore.content :
+                   '';
+
+    const suffix = afterVisible ?
+                   styleAfter.content :
+                   '';
 
     if ((prefix[0] === '"') && !prefix.toLowerCase().includes('moz-')) {
       result = prefix.substring(1, (prefix.length-1)) + result;
@@ -16850,13 +16951,13 @@
     /**
      * @method hasContent
      *
-     * @desc
+     * @desc Returns true if the DOMText has visiible content, otherwise false
      *
-     * @return {Boolean} Returns true if the DOMText has content, otherwise false
+     * @return {Boolean} see @desc
      */
 
     get hasContent () {
-      return this.text.length;
+      return this.text.length > 0;
     }
 
     addText (text) {
@@ -29365,6 +29466,7 @@
   /* resultSummary.js */
 
   const debug$I = new DebugLogging('ruleResultSummary', false);
+  debug$I.flag = false;
 
   /* ---------------------------------------------------------------- */
   /*                             RuleResultsSummary                        */
@@ -29501,6 +29603,7 @@
 
   /* Constants */
   const debug$H = new DebugLogging('ruleGroupResult', false);
+  debug$H.flag = false;
 
   /**
    * @class RuleGroupResult
@@ -29719,6 +29822,7 @@
 
   /* constants */
   const debug$G = new DebugLogging('baseResult', false);
+  debug$G.flag = false;
 
   /**
    * @class baseResult
@@ -30291,6 +30395,7 @@
   /* elementResultSummary.js */
 
   const debug$E = new DebugLogging('ElementResultSummary', false);
+  debug$E.flag = false;
 
   /* ---------------------------------------------------------------- */
   /*                             ResultSummary                        */
@@ -30433,6 +30538,7 @@
   /* Constants */
 
   const debug$D = new DebugLogging('PageResult', false);
+  debug$D.flag = false;
 
   /**
    * @class PageResult
@@ -30505,6 +30611,7 @@
   /* Constants */
 
   const debug$C = new DebugLogging('PageResult', false);
+  debug$C.flag = false;
 
   /**
    * @class WebsiteResult