+++ src/opena11y-for-toc/opena11y-for-toc.js
@@ -29,17 +29,10 @@ document.body.appendChild(scriptNode);
 // Listen for messages from side panel
 browserRuntime.onMessage.addListener(
   function(request, sender, sendResponse) {
-    debug && console.log(sender.tab ?
-                "[content.js]: from a content script:" + sender.tab.url :
-                "[content.js]: from the extension");
 
     // Highlight elements
     if(request.highlight) {
-      debug && console.log(`[content.js][highlight][position]: ${request.highlight.position}`);
-      debug && console.log(`[content.js][highlight][    info]: ${request.highlight.info}`);
-
       const he = document.querySelector(HIGHLIGHT_ELEMENT_NAME);
-      debug && console.log(`[content.js][he]: ${he}`);
 
       if (he) {
         he.setAttribute('data-attr', 'data-opena11y-id');
@@ -49,10 +42,7 @@ browserRuntime.onMessage.addListener(
 
     // Removed highlight
     if(request.removeHighlight) {
-      console.log(`[content.js][removeHighlight]`);
-
       const he = document.querySelector(HIGHLIGHT_ELEMENT_NAME);
-      console.log(`[content.js][he]: ${he}`);
 
       if (he) {
         he.setAttribute('highlight-position', '');
@@ -62,10 +52,8 @@ browserRuntime.onMessage.addListener(
     // Update Highlight configuration
     if(request.updateHighlightConfig) {
       const hc = request.updateHighlightConfig;
-      debug && console.log(`[content.js][hc]: ${hc}`);
 
       const he = document.querySelector(HIGHLIGHT_ELEMENT_NAME);
-      debug && console.log(`[content.js][he]: ${he}`);
 
       if (he) {
         he.setAttribute('highlight-config', `${hc.size} ${hc.style}`);
@@ -74,10 +62,7 @@ browserRuntime.onMessage.addListener(
 
     // Focus elements
     if(request.focusPosition) {
-      debug && console.log(`[content.js][focusElement]: ${request.focusPosition}`);
-
       const he = document.querySelector(HIGHLIGHT_ELEMENT_NAME);
-      debug && console.log(`[content.js][he]: ${he}`);
 
       if (he) {
         he.setAttribute('data-attr', 'data-opena11y-id');
@@ -97,12 +82,8 @@ browserRuntime.onMessage.addListener(
                                 '',
                                 true);
 
-      debug && console.log(`[content.js][EvalResult][   title]: ${evaluationResult.getTitle()}`);
-      debug && console.log(`[content.js][EvalResult][headings]: ${evaluationResult.headings.toJSON()}`);
-      debug && console.log(`[content.js][EvalResult][ regions]: ${evaluationResult.landmarkRegions.toJSON()}`);
-      debug && console.log(`[content.js][EvalResult][   links]: ${evaluationResult.links.toJSON()}`);
-
       sendResponse({title: evaluationResult.getTitle(),
+                    url: evaluationResult.getURL(),
                     headings: evaluationResult.headings.data,
                     regions: evaluationResult.landmarkRegions.data,
                     links: evaluationResult.links.data});
@@ -111,14 +92,20 @@ browserRuntime.onMessage.addListener(
   }
 );
 
+/*
+ * Detecting side panel closing
+ */
+
 // Check if side panel is open
+
+let openFlag = true;
 
 setInterval(() => {
   chrome.runtime
     .sendMessage({ ['toc-sidepanel-open']: true })
     .then((msgRes) => {
-      if (msgRes !== true) {
-        console.log(`TOC Sidepanel Closed`);
+      if (msgRes !== true && openFlag) {
+        openFlag = false;
       }
     })
     .catch( () => {
@@ -126,5 +113,7 @@ setInterval(() => {
       if (he) {
         he.setAttribute('highlight-position', '');
       }
-    });
-}, 100);
+      openFlag = true;
+  });
+}, 50);
+